feat(ui): appliation view page
