+++ ui/src/routes/grant-pools/[actionHashB64]/applications/+page.svelte
@@ -1,0 +1,1 @@
+VIEW ALL APPLICATIONS FOR GRANT POOL
